backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
authlib==1.3.2
httpx==0.27.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
itsdangerous==2.2.0
nh3==0.2.18
redis[hiredis]==5.1.0
slowapi==0.1.9
meilisearch==0.31.4
boto3==1.35.0
structlog==24.4.0
